requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
jiter==0.5.0
Mako==1.3.6
MarkupSafe==3.0.2
openai==1.51.0
outcome==1.3.0.post0
packaging==24.1
playwright==1.48.0
pluggy==1.5.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
PySocks==1.7.1
pytest==8.3.3
pytest-base-url==2.1.0
pytest-playwright==0.5.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
requests==2.32.3
selenium==4.25.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.38.6
text-unidecode==1.3
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
